買い物リストの全件削除を実装

+++ client/src/hooks/useShoppingItem.ts
@@ -2,51 +2,92 @@ import axios from '@/lib/axios';
 import { IGetShoppingItem, IPostShoppingItem } from '@/types/api';
 import React from 'react';
 import useSWR from 'swr';
-// import { useDebounce } from './useDebounce';
 
 const fetchShoppingItems = (path: string): Promise<IGetShoppingItem[]> =>
     axios.get(path).then(res => res.data);
 
 export const useShoppingItem = () => {
+    const [isLoading, setIsLoading] = React.useState(false);
     const {
         data: shoppingItems,
         error,
         mutate,
+        isValidating,
     } = useSWR('/api/group/shopping/items', fetchShoppingItems);
+
+    React.useEffect(() => {
+        if (isValidating) {
+            setIsLoading(true);
+        } else {
+            setIsLoading(false);
+        }
+    }, [isValidating]);
 
     const updateShoppingItems = async (items: IPostShoppingItem[]) => {
         if (
-            items.length > 0 &&
-            JSON.stringify(items) !== JSON.stringify(shoppingItems)
+            isLoading ||
+            items.length === 0 ||
+            JSON.stringify(items) === JSON.stringify(shoppingItems)
         ) {
-            try {
-                const res = await axios.post(`/api/group/shopping/items`, {
-                    items,
-                });
-                if (res.data) {
-                    // TODO: スナックバーで表示
-                    console.log(res.data.message);
-                    // 更新後にSWRのキャッシュを更新
-                    await mutate();
-                }
-            } catch (error) {
-                // TODO: スナックバーでエラー表示
-                console.error(error.response?.data.message);
-            }
+            return;
         }
-    };
 
-    const deleteShoppingItem = async (id: string) => {
         try {
-            const res = await axios.delete(`/api/group/shopping/items`, {
-                data: { id },
+            setIsLoading(true);
+            const res = await axios.post(`/api/group/shopping/items`, {
+                items,
             });
             if (res.data) {
+                // TODO: スナックバーで表示
+                console.log(res.data.message);
                 await mutate();
             }
         } catch (error) {
             // TODO: スナックバーでエラー表示
             console.error(error.response?.data.message);
+        } finally {
+            setIsLoading(false);
+        }
+    };
+
+    const deleteShoppingItem = async (id: string) => {
+        if (isLoading) return;
+
+        try {
+            setIsLoading(true);
+            const res = await axios.delete(`/api/group/shopping/items`, {
+                data: { id },
+            });
+            if (res.data) {
+                // TODO: スナックバーで表示
+                console.log(res.data.message);
+                await mutate();
+            }
+        } catch (error) {
+            // TODO: スナックバーでエラー表示
+            console.error(error.response?.data.message);
+        } finally {
+            setIsLoading(false);
+        }
+    };
+
+    const deleteAllShoppingItems = async () => {
+        if (isLoading) return;
+
+        try {
+            setIsLoading(true);
+            const res = await axios.delete(`/api/group/shopping/items/all`);
+
+            if (res.data) {
+                // TODO: スナックバーで表示
+                console.log(res.data.message);
+                await mutate();
+            }
+        } catch (error) {
+            // TODO: スナックバーでエラー表示
+            console.error(error.response?.data.message);
+        } finally {
+            setIsLoading(false);
         }
     };
 
@@ -58,8 +99,10 @@ export const useShoppingItem = () => {
     }, [error]);
 
     return {
+        isLoading,
         shoppingItems,
         updateShoppingItems,
         deleteShoppingItem,
+        deleteAllShoppingItems,
     };
 };
